Move rendering into index.ts and make  AnimationCotnrol  a class.
